1.1.4 (#307)
* Reference fix

* Added EmotionPresetType wrapper

* Replaced MaxSergeants and MaxCaptains with SergeantsPercentage and CaptainsPercentage

* gitlab templates

* README changes

* feat: new events

* docs: better spacing and langwording

* fix: event args naming scheme

* docs: fix inconsistencies

* Scp173 Snap Events

* Apply 1 suggestion(s) to 1 file(s)


* Add CheckedHitmarker event
Add Send/ing Hitmarker event
Add log for new events.

* Fix docs.
SendHitmarker to SentHitmarker

* Add Scp049 Attacked / ing
Add Scp049 Sense Lost/ Killed Target

* Apply 4 suggestion(s) to 2 file(s)


* Apply 10 suggestion(s) to 5 file(s)


* fix IsAllowed not set in Attacking
Added log of the new events

* Apply suggestions

* Rename PlayAudio to PlayedAudio
Revert LogPlugin handlers tolog everything too

* Add Wave event

* Apply 5 suggestion(s) to 2 file(s)


* Fix nullable target

* basic required version check

* allow force-loading unsupported versions + improve messages

* Add Focuesd
Add Mimiced/ing Enviroment

* version bump just in case

* Fix methods

* per-port global LoadUnsupportedPlugins property

* rename ForceUnsupportedMajorVersion

* OutdatedLoadingBehavior

* Adding support for multiple command handlers.

* make OutdatedLoadingBehavior more generic & reorder its values

* Apply 13 suggestion(s) to 4 file(s)


* Rename Mimicing -> Mimicking
Rename Enviroment -> Environment

* Fix event name in xml
ReAdd events

* Add Map events

* Push changes (#282)

* removed line breaks from the description of UnsupportedLoading in Properties

* Version null check

* Fix build & revision comparing

---------

+++ LabApi/Features/Wrappers/Facility/Warhead.cs
@@ -347,7 +347,8 @@ public static class Warhead
         internal DetonationScenario(AlphaWarheadController.DetonationScenario detonationScenario, byte id, WarheadScenarioType type)
         {
             TimeToDetonate = detonationScenario.TimeToDetonate;
-            AdditionalTime = detonationScenario.AdditionalTime;
+            // TODO: Remove the cast on 2.0.0 and change additional time to float
+            AdditionalTime = (int)detonationScenario.AdditionalTime;
             Type = type;
             Id = id;
         }
